@@ -3,34 +3,27 @@ package org.selfconference.android.utils.rx;
 import android.support.v4.widget.SwipeRefreshLayout;
 import rx.Observable;
 import rx.Observable.Transformer;
-import rx.functions.Action0;
+import rx.android.schedulers.AndroidSchedulers;
+import rx.schedulers.Schedulers;
 
 public final class Transformers {
 
   public static <T> Transformer<T, T> setRefreshing(final SwipeRefreshLayout swipeRefreshLayout) {
-    return new Transformer<T, T>() {
-      @Override public Observable<T> call(Observable<T> observable) {
-        return observable //
-            .doOnSubscribe(new Action0() {
-              @Override public void call() {
-                swipeRefreshLayout.post(new Runnable() {
-                  @Override public void run() {
-                    swipeRefreshLayout.setRefreshing(true);
-                  }
-                });
-              }
-            }) //
-            .doOnTerminate(new Action0() {
-              @Override public void call() {
-                swipeRefreshLayout.post(new Runnable() {
-                  @Override public void run() {
-                    swipeRefreshLayout.setRefreshing(false);
-                  }
-                });
-              }
-            });
-      }
-    };
+    return observable -> observable //
+        .doOnSubscribe(
+            () -> swipeRefreshLayout.post(() -> swipeRefreshLayout.setRefreshing(true))) //
+        .doOnTerminate(
+            () -> swipeRefreshLayout.post(() -> swipeRefreshLayout.setRefreshing(false)));
+  }
+
+  /**
+   * Composes an {@link Observable} to subscribe an I/O thread scheduler
+   * and observe the result on the UI thread.
+   */
+  public static <T> Transformer<T, T> ioSchedulers() {
+    return observable -> observable //
+        .subscribeOn(Schedulers.io()) //
+        .observeOn(AndroidSchedulers.mainThread());
   }
 
   private Transformers() {
