@@ -2,17 +2,13 @@ package org.selfconference.android;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import butterknife.ButterKnife;
-import rx.Subscription;
-import rx.subscriptions.CompositeSubscription;
+import com.trello.rxlifecycle.components.support.RxFragment;
 
-public abstract class BaseFragment extends Fragment {
-
-  private final CompositeSubscription compositeSubscription = new CompositeSubscription();
+public abstract class BaseFragment extends RxFragment {
 
   @Override public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container,
       @Nullable Bundle savedInstanceState) {
@@ -21,21 +17,12 @@ public abstract class BaseFragment extends Fragment {
 
   @Override public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
     super.onViewCreated(view, savedInstanceState);
-    ButterKnife.inject(this, view);
+    ButterKnife.bind(this, view);
   }
 
   @Override public void onActivityCreated(@Nullable Bundle savedInstanceState) {
     super.onActivityCreated(savedInstanceState);
     App.getInstance().inject(this);
-  }
-
-  @Override public void onDestroy() {
-    super.onDestroy();
-    compositeSubscription.unsubscribe();
-  }
-
-  protected void addSubscription(Subscription subscription) {
-    compositeSubscription.add(subscription);
   }
 
   protected abstract int layoutResId();
