@@ -5,14 +5,14 @@ import android.support.annotation.Nullable;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.view.View;
-import butterknife.InjectView;
+import butterknife.Bind;
 import org.selfconference.android.BaseFragment;
 import org.selfconference.android.R;
 
 public class CodeOfConductFragment extends BaseFragment {
   public static final String TAG = CodeOfConductFragment.class.getName();
 
-  @InjectView(R.id.code_of_conduct_recycler_view) RecyclerView recyclerView;
+  @Bind(R.id.code_of_conduct_recycler_view) RecyclerView recyclerView;
 
   @Override protected int layoutResId() {
     return R.layout.fragment_code_of_conduct;
