@@ -5,8 +5,8 @@ import android.support.annotation.Nullable;
 import android.support.v4.widget.SwipeRefreshLayout;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import butterknife.InjectView;
-import java.util.Collections;
+import butterknife.Bind;
+import com.google.common.collect.Ordering;
 import java.util.List;
 import javax.inject.Inject;
 import org.selfconference.android.BaseListFragment;
@@ -16,19 +16,17 @@ import org.selfconference.android.api.Api;
 import org.selfconference.android.api.ApiRequestSubscriber;
 import org.selfconference.android.sponsors.SponsorAdapter.OnSponsorClickListener;
 import org.selfconference.android.utils.Intents;
-import org.selfconference.android.utils.rx.Transformers;
-import rx.Observable;
 import rx.Subscriber;
-import rx.functions.Func1;
 import timber.log.Timber;
 
-import static rx.android.app.AppObservable.bindFragment;
+import static org.selfconference.android.utils.rx.Transformers.ioSchedulers;
+import static org.selfconference.android.utils.rx.Transformers.setRefreshing;
 
 public class SponsorListFragment extends BaseListFragment implements OnSponsorClickListener {
   public static final String TAG = SponsorListFragment.class.getName();
 
-  @InjectView(R.id.sponsor_recycler_view) RecyclerView sponsorRecyclerView;
-  @InjectView(R.id.sponsor_list_swipe_refresh_layout) SwipeRefreshLayout swipeRefreshLayout;
+  @Bind(R.id.sponsor_recycler_view) RecyclerView sponsorRecyclerView;
+  @Bind(R.id.sponsor_list_swipe_refresh_layout) SwipeRefreshLayout swipeRefreshLayout;
 
   @Inject Api api;
 
@@ -45,16 +43,12 @@ public class SponsorListFragment extends BaseListFragment implements OnSponsorCl
     sponsorRecyclerView.setAdapter(sponsorAdapter);
     sponsorRecyclerView.setLayoutManager(new LinearLayoutManager(getActivity()));
 
-    final Observable<List<Sponsor>> sponsorsObservable = bindFragment(this, api.getSponsors()) //
-        .map(new Func1<List<Sponsor>, List<Sponsor>>() {
-          @Override public List<Sponsor> call(List<Sponsor> sponsors) {
-            Collections.sort(sponsors, new SponsorComparator());
-            return sponsors;
-          }
-        }) //
-        .compose(Transformers.<List<Sponsor>>setRefreshing(swipeRefreshLayout));
-
-    addSubscription(sponsorsObservable.subscribe(sponsorsSubscriber));
+    api.getSponsors() //
+        .map(sponsors -> Ordering.from(new SponsorComparator()).immutableSortedCopy(sponsors)) //
+        .compose(setRefreshing(swipeRefreshLayout))
+        .compose(bindToLifecycle())
+        .compose(ioSchedulers())
+        .subscribe(sponsorsSubscriber);
   }
 
   @Override protected FilterableAdapter getFilterableAdapter() {
