@@ -7,13 +7,13 @@ import org.selfconference.android.Parceler.Container;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.selfconference.android.Parceler.testParceling;
-import static org.selfconference.android.feedback.Vote.NEGATIVE;
-import static org.selfconference.android.feedback.Vote.POSITIVE;
+import static org.selfconference.android.feedback.VoteButton.VOTE_NEGATIVE;
+import static org.selfconference.android.feedback.VoteButton.VOTE_POSITIVE;
 
 @RunWith(CustomTestRunner.class) public class FeedbackTest {
 
   @Test public void parcelsNegativeFeedbackWithoutError() {
-    final Feedback feedback = new Feedback(NEGATIVE, "test comments");
+    final Feedback feedback = new Feedback(VOTE_NEGATIVE, "test comments");
 
     final Container<Feedback> feedbackContainer = testParceling(feedback, Feedback.CREATOR);
 
@@ -21,7 +21,7 @@ import static org.selfconference.android.feedback.Vote.POSITIVE;
   }
 
   @Test public void parcelsPositiveFeedbackWithoutError() {
-    final Feedback feedback = new Feedback(POSITIVE, "test comments");
+    final Feedback feedback = new Feedback(VOTE_POSITIVE, "test comments");
 
     final Container<Feedback> feedbackContainer = testParceling(feedback, Feedback.CREATOR);
 
